--- src/main/java/FileManager.java
@@ -1,2 +1,0 @@
-public class FileManager {
-}
