+++ src/main/java/duke/Deadline.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Deadline extends Task {
     protected final String by;
