+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Task {
     protected final String description;
     protected boolean isDone;
